ready to work

+++ ft_atoi.c
@@ -6,15 +6,17 @@
 /*   By: yallo <yallo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/09 13:35:14 by yallo             #+#    #+#             */
-/*   Updated: 2022/11/18 11:49:34 by yallo            ###   ########.fr       */
+/*   Updated: 2022/11/22 13:06:45 by yallo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 int	ft_atoi(const char *str)
 {
-	int	i;
-	int	result;
-	int	sign;
+	size_t	i;
+	int		result;
+	int		sign;
 
 	i = 0;
 	sign = 1;
